% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths.club-package.R
\docType{package}
\name{maths.club-package}
\alias{maths.club}
\alias{maths.club-package}
\title{maths.club: Examples for maths club}
\description{
Contains some examples for maths club.
}
\author{
\strong{Maintainer}: Rob Ashton \email{r.ashton@imperial.ac.uk}

}
\keyword{internal}
