% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis_hastings.R
\name{metropolis}
\alias{metropolis}
\title{Run metropolis hastings MCMC}
\usage{
metropolis(f, initial, steps, propsal_sd)
}
\arguments{
\item{propsal_sd}{}
}
\value{

}
\description{
Run metropolis hastings MCMC
}
